fix(maptos-opt-executor): decouple indexer
Don't return the indexer runtime as a return value
from Executor::background. It ends up unused in the tests
and dropped early in some of them, causing a crash.

Context::run_indexer_grpc_service is a perfectly good way
to set up and run the indexer when it is needed.