fix(cicd): no symbolic references in checkout
Don't supply the symbolic GitHub reference name to the `ref` parameter
of the checkout action in the build-push-container-all workflow jobs.
The default behavior of actions/checkout would do the right thing
checking out the SHA of the triggering event, thereby
avoiding races with concurrent pushes to the repository modifying
the symbolic references.