refactor(signer): simplify the signing API
Remove the unnecessary Signer and Verifier structs.
Rename the *Operations traits to Signer and Verify because Rust.
Remove use of async_trait: we don't want to force boxing on all
implementations. Use impl Future in return signatures instead.
The Verify trait is made simply synchronous because it works with
immediately available data and can be implemented without network calls.