Convert DER signature to raw signature for AWS KMS signing (#988)
For now, given Milestone 1 relating to auth rather than key standards, I've changed this PR's purpose to simply adding the DER to raw signature conversion to make signing work with an ARN Key ID as an env var. I'll use @l-monninger and @musitdev's comments on this PR as guidance when we get to integrating Parameter Store for signing.