refactor!: YAGNIfy prevalidator
Remove heaps of abstraction and whittle down the transaction
prevalidator to what it really needs to do in the DA light node.
No wasteful back-and-forth serializing back to movement Transaction
struct to fulfill an abstract transformation-like trait. No async.

Also we get to use the same Validator instance in the case when
no whitelist is given, instead of bypassing the whole thing as
the light node sequencer module previously did, opening the DA to
arbitrarily malformed transactions.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch mikhail/prevalidator-yagni
# You are currently bisecting, started from branch 'main'.
#
# Changes to be committed:
#	modified:   protocol-units/da/movement/protocol/light-node/src/sequencer.rs
#	new file:   protocol-units/da/movement/protocol/prevalidator/src/aptos.rs
#	deleted:    protocol-units/da/movement/protocol/prevalidator/src/aptos/mod.rs
#	deleted:    protocol-units/da/movement/protocol/prevalidator/src/aptos/transaction.rs
#	deleted:    protocol-units/da/movement/protocol/prevalidator/src/aptos/whitelist/mod.rs
#	deleted:    protocol-units/da/movement/protocol/prevalidator/src/aptos/whitelist/whitelist.rs
#	modified:   protocol-units/da/movement/protocol/prevalidator/src/lib.rs
#